fix: resolve 6 failing CI tests for string concat and class type inference
Fixes two root causes of CI test failures:

1. Class Type Inference (test_46, test_50):
   - When HM inference produces UnificationVar for class constructors,
     the type mapper was falling back to HashSet<DepylerValue>
   - Added detection for class constructor calls with unresolved types
   - Now generates correct type annotation: `let calc: Calculator = ...`

2. String Concatenation Cast (test_03, test_36, test_38, test_138):
   - Extended expr_is_string_type to handle Binary Add expressions
   - Added target_is_string check to prevent invalid `as i32` casts
   - String concatenation chains now handled correctly

All 6 previously failing tests now pass:
- test_03_literals_strings
- test_36_string_methods
- test_38_string_formatting
- test_46_instance_method
- test_50_method_chaining
- test_138_string_concatenation

+++ examples/debugging_workflow.rs
@@ -3111,7 +3111,7 @@ pub fn fibonacci(n: i32) -> i32 {
     if _cse_temp_0 {
         return n;
     }
-    (fibonacci((n).py_sub(1))).py_add(fibonacci((n).py_sub(2)))
+    (fibonacci((n).py_sub(1i32))).py_add(fibonacci((n).py_sub(2i32)))
 }
 #[doc = "Calculate factorial iteratively"]
 #[doc = " Depyler: verified panic-free"]
@@ -3119,8 +3119,8 @@ pub fn fibonacci(n: i32) -> i32 {
 pub fn factorial(n: i32) -> i32 {
     let mut result: i32 = Default::default();
     result = 1;
-    for i in (2)..((n).py_add(1)) {
-        result = (result).py_mul(i);
+    for i in (2)..((n).py_add(1i32)) {
+        result = ((result).py_mul(i)) as i32;
     }
     result
 }
